move pgx specific code from datastore/common to datastore/postgres/common
